show_expense_dashboard

+++ controllers/expense.js
@@ -1,4 +1,5 @@
 const Expense = require('../models/expense');
+const User = require('../models/user');
 
 function stringInvalid(string) {
     if( string == undefined || string.length === 0 )
@@ -9,7 +10,8 @@ function stringInvalid(string) {
 
 exports.getExpenses = async(req, res, next) => {
     try {
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({ where: {userId: req.user.id}});
+       //const expenses = await req.user.getExpenses();
         res.status(200).json({allExpenses: expenses});;
 
     } catch(error) {
@@ -23,12 +25,14 @@ exports.postExpense = async (req, res, next) => {
         const amount = req.body.amount;
         const desc = req.body.desc;
         const category = req.body.category;
+        const userId = req.user.id;
 
         if(amount === undefined || stringInvalid(desc) || stringInvalid(category)) {
             return res.status(400).json({success: false, message:'Input missing'});
         }
+         //req.user.createExpenses=( {amount: amount, desc: desc, category: category});
 
-        const data = await Expense.create( {amount: amount, desc: desc, category: category});
+       const data = await Expense.create( {amount: amount, desc: desc, category: category, userId: userId});
         res.status(201).json({newExpenseDetail: data, success: true});
     } catch(err) {
         res.status(500).json({error: err, success: false})
@@ -43,10 +47,12 @@ exports.deleteExpense = async(req, res, next) => {
             return res.status(404).json({err: 'Id is missing'});
         }
         const expenseId = req.params.id;
-        await Expense.destroy({where: {id: expenseId}});
+        const noOfRows = await Expense.destroy({where: {id: expenseId, userId: req.user.id}});
+        if(noOfRows === 0)
+            return res.status(404).json({success: false, message: 'Expense does not belong to the user'});
         res.sendStatus(200);
     } catch(error) {
         console.log('Delete user is failing '+ JSON.stringify(error));
-        res.status(500).json({ error: error});
+        res.status(500).json({success: false, message: 'Failed'});
     }
 };
